Add files via upload
As an idea for my first project, I thought of analyzing data to understand which tourist destinations are the most popular and why, by selecting sample cities and analyzing various aspects, including the average spending on accommodation, rather than the main attractions of the place. I used MySQL for the data cleaning and modeling process, and finally created the data visualization in Power BI. The dataset used in this project was generated by ChatGPT. It was created for the purpose of this project and may not reflect real-world data.